+++ lib/interfaces/type-error.ts
@@ -2,6 +2,7 @@ import {
   AuthError,
   CustomerError,
   CustomError,
+  ForbiddenError,
   NotFoundError,
   SchemaError,
 } from '../utilities/errors';
@@ -12,6 +13,7 @@ export type ErrorType =
   | CustomerError
   | NotFoundError
   | AuthError
+  | ForbiddenError
   | CustomError
   | Record<string, any>;
 
@@ -21,5 +23,6 @@ export enum ErrorIdentifier {
   NOT_FOUND = 'not_found',
   INTERNAL = 'internal',
   UNAUTHORIZED = 'unauthorized',
+  FORBIDDEN = 'forbidden',
   CUSTOM = 'custom',
 }